Official and member reviews
Code for scrapping official and member reviews from beyazperde.com.

The folders also contain scrapped official and member reviews, updated as of 15/15/2021 and 9/12/2021 respectively.

Group reviews to single file folder has the code for reading from the compiled files and creates new excel files for each movie with each review on a separate worksheet with the workbooks.